requirements.txt
flask
openai-whisper
#imageio[ffmpeg]
ffmpeg-binaries